nyhende-server/src/main/java/no/plasmid/nyhende/controller/HomeController.java
package no.plasmid.nyhende.controller;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
@RequestMapping("/")
public class HomeController {

    private static final Logger LOG = LoggerFactory.getLogger(HomeController.class);

    @GetMapping
    public String getIndex(Model model) {
        LOG.debug("Get webpage index.html");
        return "forward:/index.html";
    }

}
